YACSA-7 Clean-up code
