+++ core/ui/src/main/kotlin/dev/yacsa/ui/window/RememberWindowInfo.kt
@@ -2,7 +2,6 @@ package dev.yacsa.ui.window
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 
 @Composable
@@ -22,17 +21,4 @@ fun rememberWindowInfo(): WindowInfo {
         screenWidth = configuration.screenWidthDp.dp,
         screenHeight = configuration.screenHeightDp.dp,
     )
-}
-
-data class WindowInfo(
-    val screenWidthInfo: WindowType,
-    val screenHeightInfo: WindowType,
-    val screenWidth: Dp,
-    val screenHeight: Dp,
-) {
-    sealed class WindowType {
-        object Compact : WindowType()
-        object Medium : WindowType()
-        object Expanded : WindowType()
-    }
 }
