+++ core/ui/src/main/kotlin/dev/yacsa/ui/window/WindowInfo.kt
@@ -1,0 +1,16 @@
+package dev.yacsa.ui.window
+
+import androidx.compose.ui.unit.Dp
+
+data class WindowInfo(
+    val screenWidthInfo: WindowType,
+    val screenHeightInfo: WindowType,
+    val screenWidth: Dp,
+    val screenHeight: Dp,
+) {
+    sealed class WindowType {
+        object Compact : WindowType()
+        object Medium : WindowType()
+        object Expanded : WindowType()
+    }
+}
