+++ feature/main/src/main/kotlin/dev/yacsa/main/MainViewModel.kt
@@ -2,5 +2,4 @@ package dev.yacsa.main
 
 import androidx.lifecycle.ViewModel
 
-class MainViewModel : ViewModel() {
-}
+class MainViewModel : ViewModel()
